Title: 앞에서부터 삭제하기 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240904/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -32,7 +32,7 @@ public class Main {
 
             total += arr[i];
 
-            maxResult = Math.max(maxResult, (total - pq.peek()) / (pq.size() - 1));
+            maxResult = Math.max(maxResult, (double) (total - pq.peek()) / (pq.size() - 1));
         }
         System.out.printf("%.2f", maxResult);
     }
